Fixes issue that was introduced when not appling the template. Adds the ability to change the tag and apply that to the BuildConfig object

+++ Jenkinsfile
@@ -50,6 +50,7 @@ pipeline {
             steps {
                 script {
                     def openShiftApplyArgs = ""
+                    updateBuildConfig = false
 
                     /* Read the version but make sure there is not any characters we do not expect */
                     env.VERSION = readFile('app/VERSION').trim()
@@ -68,13 +69,17 @@ pipeline {
                              * findFileChanges method is implemented in vars/findFileChanges.groovy
                              */
 
+
                             if (findFileChanges(params.APP_TEMPLATE_PATH)
                                     || !openshift.selector("template/${params.APP_DC_NAME}").exists()) {
 
                                 /* Apply the template object from JSON file */
                                 openshift.apply(readFile(params.APP_TEMPLATE_PATH))
+                                echo ">>>>> OpenShift Template Changed <<<<<"
                             } else {
                                 openShiftApplyArgs = "--dry-run"
+                                updateBuildConfig = true
+                                echo ">>>>> OpenShift Template Unchanged <<<<<"
                             }
 
                             /* Process the template and return the Map of the result */
@@ -85,7 +90,6 @@ pipeline {
                                     "IMAGESTREAM_TAG=${params.IMAGE_STREAM_LATEST_TAG}",
                                     "REGISTRY_PROJECT=${params.REGISTRY_PROJECT}",
                                     "REGISTRY=${params.REGISTRY_URI}")
-
 
                            	/* If the project's secret exists it must be removed at least in
                              * the case of using a database along with an application.
@@ -108,7 +112,9 @@ pipeline {
                                 }
                             }
 
+                            echo ">>>>> Enter OpenShift Apply Template Model <<<<<"
                             createdObjects = openshift.apply(model, openShiftApplyArgs)
+                            echo ">>>>> Exit OpenShift Apply Template Model <<<<<"
                         }
                     }
                 }
@@ -128,7 +134,18 @@ pipeline {
                             /* Select only BuildConfig from the objects that were created from the template
                              * process and apply
                              */
+                            def buildConfig = openshift.selector("buildconfig/${params.APP_DC_NAME}").object()
                             buildConfigs = createdObjects.narrow('bc')
+                            imageName = "${((String) buildConfig.spec.output.to.name).split(':')[0]}:${env.TAG}"
+
+                            if (updateBuildConfig) {
+                                echo ">>>>> Enter OpenShift Apply BuildConfig <<<<<"
+                                buildConfig.spec.output.to.name = imageName
+                                openshift.apply(buildConfig)
+                                openshift.raw("annotate", "buildconfig/${params.APP_DC_NAME}",
+                                        "kubectl.kubernetes.io/last-applied-configuration-")
+                                echo ">>>>> Exit OpenShift Apply BuildConfig <<<<<"
+                            }
 
                             def build = null
                             def buildErrorLog = "Build Error"
@@ -145,10 +162,9 @@ pipeline {
                                 }
                                 build.untilEach {
                                     /* Wait until a build object is complete */
+
                                     echo ">>>>> Enter Build Status Check <<<<<"
-
                                     def phase = it.object().status.phase
-
                                     echo ">>>>> Exit Build Status Check - Phase: ${phase} <<<<<"
 
                                     if (phase == "Complete") {
@@ -171,6 +187,10 @@ pipeline {
                                 return
                             }
 
+                            openshift.tag("--source=docker",
+                                    "${imageName}",
+                                    "${openshift.project()}/${params.IMAGE_STREAM_NAME}:${params.IMAGE_STREAM_LATEST_TAG}")
+
                             /* Resolves issue with an ImageStream that won't initially
                              * import a tag
                              *
@@ -187,7 +207,7 @@ pipeline {
                             timeout(10) {
                                 createdObjects.narrow('is').untilEach {
                                     echo ">>>>> Enter ImageStream Check <<<<<"
-                                    echo "importImage: ${importImage} reCheck: ${reCheck}"
+                                    echo ">>>>> importImage: ${importImage} reCheck: ${reCheck} <<<<<"
 
                                     if (importImage) {
                                         reCheck = true
@@ -201,7 +221,7 @@ pipeline {
                                         }
                                     }
                                     echo ">>>>> Exit ImageStream Check <<<<<"
-                                    echo "importImage: ${importImage} reCheck: ${reCheck}"
+                                    echo ">>>>> importImage: ${importImage} reCheck: ${reCheck} <<<<<"
 
                                     /* NXOR - if importImage and reCheck are both true or false
                                      * return true, else return false
@@ -258,7 +278,6 @@ pipeline {
          * This is generally duplicated from the Dev - OpenShift Template
          * See inline documentation above.
          */
-
         stage('Stage - OpenShift Template') {
             environment {
                 STAGE = credentials("${params.STAGE_SECRET_NAME}")
@@ -273,6 +292,9 @@ pipeline {
                                 openshift.apply(readFile(params.APP_TEMPLATE_PATH))
                             } else {
                                 openShiftApplyArgs = "--dry-run"
+                                openshift.tag("--source=docker",
+                                        "${imageName}",
+                                        "${openshift.project()}/${params.IMAGE_STREAM_NAME}:${params.IMAGE_STREAM_LATEST_TAG}")
                             }
 
                             def model = openshift.process(params.IMAGE_STREAM_NAME,
@@ -295,7 +317,9 @@ pipeline {
                             createdObjects = openshift.apply(model, openShiftApplyArgs)
 
                             /* The stage environment does not need OpenShift BuildConfig objects */
-                            createdObjects.narrow('bc').delete()
+                            if (createdObjects.narrow('bc').exists()) {
+                                createdObjects.narrow('bc').delete()
+                            }
                         }
                     }
                 }
